backend/requirements.txt
# ============================================
# Core Framework
# ============================================
aiogram==3.4.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================
# Database
# ============================================
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# ============================================
# Background Tasks
# ============================================
celery==5.3.6
flower==2.0.1

# ============================================
# Validation & Serialization
# ============================================
pydantic==2.5.3
pydantic-settings==2.1.0
marshmallow==3.20.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# ============================================
# HTTP Client
# ============================================
aiohttp==3.9.1
httpx==0.26.0

# ============================================
# Image Processing
# ============================================
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# ============================================
# S3 Storage
# ============================================
minio==7.2.0
aiobotocore==2.11.0

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
orjson==3.9.10
structlog==24.1.0
 tenacity==8.2.3

# ============================================
# Monitoring
# ============================================
sentry-sdk==1.39.2
prometheus-client==0.19.0

# ============================================
# Testing
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# ============================================
# Development
# ============================================
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
